load peggy parsers directly rather than pregenerating
Instead of having a build step to convert the pegjs files into js files,
load them directly. This requires creating a module loader to be convert
the pegjs files as they are loaded. This makes development a bit
simpler.

The released form of this is a single file compiled with esbuild. Add a
plugin to esbuild that converts the pegjs files as well, so the compiled
form doesn't need the peggy parser generator.